feat!: create resume header section
Remove Ant Design and change the UI Library to MUI because the
documentation is easier to understand for a simple project such
as Resumetor. Overhauled components that uses the UI made for this
project to integrate the component libraries from MUI.

Create utils folder for the makeArrayHeaderData and shortenLink
function. The makeArrayHeaderData function will turn the data from the
useForm of react-hook-form to an array of objects containing a
key, the name of the input, and value, the content of the input field.
If the key is a link, the value will be an object containing
a url that has a value of the original link, and a shorthand that holds
the short url version of the link. The shortenLink function converts a
url string to a shorter version. This function returns an object with
the url and shorthand property.

Modified the Section and SectionTitle component to convert all of its
jsx element to a react MUI element.

Create an interface folder to contain the interface used by hooks and
functions related to react-hook-form. This change will help standardize
the use of types for different values utilized by the react-hook-form
library.

BREAKING CHANGE: use components in MUI library not installed before
Refs: #11