feat: add EducationSubsection to EducationSection component
Create EducationSubsection to easily work with nested field arrays with
react-hook-form. Working with nested field arrays with react-hook-form
is challenging because you need to find the index of the parent field
array to make append work properly. Separting the concern of a nested
field array will make indexing easier for the useFieldArray() and easier
to maintain.